Tightens lifecycle and finishes wazero
The previous commit didn't work because wazero didn't implement the ABI
registry functionality in the wasmer engine. This fixes that.

This also adds the ability to close a VM, which helps from leaking
resources.

Finally, the backports the tests that were in wasmer and relevant to
wazero. I tested this with mosn which now can run the examples etc.

Signed-off-by: Adrian Cole <adrian@tetrate.io>